requirements.txt
backoff
beautifulsoup4
cryptography
datetime
ddgs
docutils
HTMLParser
ipython
ipywidgets
logging
js
json5
numpy
openai
pandas
playwright
pyOpenSSL
pymongo
python-dotenv
python_dateutil
readability_lxml
redis
requests
simplejson
thread
tqdm
typing_extensions
zstandard